refactor: utils from services
