+++ packages/ui/src/components/va-alert/VaAlert.vue
@@ -81,7 +81,7 @@
 <script lang="ts">
 import { defineComponent, computed, PropType } from 'vue'
 
-import { generateUniqueId } from '../../services/utils'
+import { generateUniqueId } from '../../utils/uuid'
 import {
   useComponentPresetProp,
   useStateful, useStatefulProps, useStatefulEmits,
