+++ packages/ui/src/composables/index.ts
@@ -59,3 +59,4 @@ export * from './useTrapFocus'
 export * from './useValidation'
 export * from './useWindow'
 export * from './useWindowSize'
+export * from './useSlotPassed'
