+++ packages/ui/src/utils/sleep.ts
@@ -1,0 +1,3 @@
+export const sleep = (ms = 0) => {
+  return new Promise(resolve => setTimeout(resolve, ms))
+}
