+++ packages/ui/src/utils/uuid.ts
@@ -1,0 +1,7 @@
+const getRandomString = (stringLength = 4): string => {
+  return Math.random().toString(36).substring(2, stringLength + 2)
+}
+
+export const generateUniqueId = () => {
+  return `${getRandomString(8)}-${getRandomString(4)}-${getRandomString(4)}`
+}
